increase swipe resolution with point-point interpolation
it was possible to swipe across the board, over a squash, and not
actually slice the squash.  This was due to the interval at which
mousemove/touchmove events fire in the browser.  The improvement takes
the current and last points from mousemove/touchmove events,
interpolates additional points along the line between them, and then
slices the veggies at those points.  Performance is a concern here but
so far it has not proven to be a problem.